bridge cluster tests into cargo test (#3592)
